requirements.txt
langchain
langchain-google-genai
langgraph
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
